backend/fastapi/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
fastapi==0.121.1
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
httptools==0.7.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
motor==3.7.1
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
python-multipart==0.0.9
cloudinary==1.41.0